# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Hades API
  description: |
    Hades 代理内核 RESTful API 规范。

    所有需要认证的接口均通过 `Authorization: Bearer <secret>` 请求头或 `?token=<secret>` 查询参数传递。
  version: 0.5.0
  contact:
    name: Qing060325
    url: https://github.com/Qing060325/Hades
servers:
  - url: http://localhost:9090
    description: 本地开发
security:
  - BearerAuth: []
paths:
...
  /config:
    get:
      summary: 获取当前配置
      operationId: getConfig
      responses:
        "200":
          description: 当前运行配置
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuntimeConfig"
    patch:
      summary: 更新部分配置
      operationId: patchConfig
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: 配置已更新
  /configs:
    put:
      summary: 上传完整配置并热重载
      description: |
        上传完整 YAML/JSON 配置并触发热重载。
        仅更新代理、代理组和规则，不重启监听器。
      operationId: putConfigs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [path]
              properties:
                path:
                  type: string
                  description: 配置文件路径
          text/yaml:
            schema:
              type: string
      responses:
        "200":
          description: 配置重载成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: config reloaded successfully
        "400":
          description: 配置重载失败
    patch:
      summary: 从文件重载配置
      operationId: patchConfigs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [path]
              properties:
                path:
                  type: string
                  description: 配置文件路径
                  example: /root/.config/hades/config.yaml
      responses:
        "200":
          description: 配置重载成功
...
  /dns/query:
    get:
      summary: DNS 查询
      operationId: queryDNS
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
          description: 要查询的域名
      responses:
        "200":
          description: 查询结果
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                  ips:
                    type: array
                    items:
                      type: string
                  error:
                    type: string
                    nullable: true
  /logs:
    get:
      summary: 获取日志流（SSE）
      operationId: getLogs
      parameters:
        - name: level
          in: query
          schema:
            type: string
            enum: [silent, error, warning, info, debug]
            default: info
      responses:
        "200":
          description: SSE 日志流
          content:
            text/event-stream:
              schema:
                type: object
...
  /subscriptions/{name}:
    get:
      summary: 获取订阅详情
      operationId: getSubscription
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 订阅详情
        "404":
          description: 订阅不存在
    put:
      summary: 更新指定订阅
      operationId: updateSubscription
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 更新成功
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Hades API 密钥
  schemas:
...
    RuntimeConfig:
      type: object
      properties:
        mode:
          type: string
          enum: [rule, global, direct]
          example: rule
        logLevel:
          type: string
          enum: [silent, error, warning, info, debug]
          example: info
        tun:
          type: object
          properties:
            enable:
              type: boolean
...